requirements.txt
# On Colab: torch/torchaudio are pre-installed — pip will skip them.
# Locally: these will install compatible versions.
torch>=2.0.0
torchaudio>=2.0.0
transformers==4.28.1
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
wandb
